+++ pkg/mimirtool/commands/analyse.go
@@ -67,8 +67,15 @@ func (cmd *AnalyzeCommand) Register(app *kingpin.Application, envVars EnvVarName
 	grafanaAnalyzeCmd.Flag("output", "The path for the output file").
 		Default("metrics-in-grafana.json").
 		StringVar(&gaCmd.outputFile)
+	grafanaAnalyzeCmd.Flag("datasource-uid", "Filter metrics by datasource uid, if defined").
+		Default("").
+		StringVar(&gaCmd.datasourceUID)
 	grafanaAnalyzeCmd.Flag("folder-title", "Limit dashboards analysis for unused metrics based on their exact folder title. When repeated any of the matching folders will be analyzed.").
 		SetValue(&gaCmd.folders)
+	grafanaAnalyzeCmd.Flag("folder-uid", "Limit dashboards analysis for unused metrics based on their folder uid. When repeated any of the matching folders will be analyzed.").
+		SetValue(&gaCmd.folderUIDs)
+	grafanaAnalyzeCmd.Flag("dashboard-id", "Limit dashboards analysis for unused metrics based on their dashboard id. When repeated any of the matching dashboards will be analyzed.").
+		SetValue(&gaCmd.dashboardIDs)
 
 	raCmd := &RulerAnalyzeCommand{}
 	rulerAnalyzeCmd := analyzeCmd.Command("ruler", "Analyze and extract the metrics that are used in Grafana Mimir rules").
@@ -97,6 +104,9 @@ func (cmd *AnalyzeCommand) Register(app *kingpin.Application, envVars EnvVarName
 	dashboardAnalyzeCmd.Flag("output", "The path for the output file").
 		Default("metrics-in-grafana.json").
 		StringVar(&daCmd.outputFile)
+	dashboardAnalyzeCmd.Flag("datasource-uid", "Filter metrics by datasource uid, if defined").
+		Default("").
+		StringVar(&daCmd.datasourceUID)
 
 	rfCmd := &RuleFileAnalyzeCommand{}
 	ruleFileAnalyzeCmd := analyzeCmd.Command("rule-file", "Analyze and output the metrics used in Prometheus rules files").Action(rfCmd.run)